pyproject.toml
[build-system]
requires = ['setuptools>=68.0', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
name = 'graph-augmented-retrieval-system'
version = '0.1.0'
description = 'A system for augmenting standard vector search with graphs for facilitating complex queries that require multiple hops across documents that seem unrelated or distant in the vector space.'
readme = 'README.md'
requires-python = '>=3.10'
authors = [{ name = 'Syed Mohammad Abdullah Rashid' }]

dependencies = [
    'wandb>=0.16',
    'pyyaml>=6.0',
    'tqdm>=4.66',
    'sentence-transformers>=2.2',
    'python-dotenv>=1.0',
    'faiss-cpu>=1.7',
    'torch>=2.0',
    'torch_geometric>=2.5',
    'spacy>=3.7',
    'scikit-learn>=1.3',
    'numpy>=1.24',
    'pandas>=2.0',
    'matplotlib>=3.7',
    'seaborn>=0.12',
    'datasets>=2.16.0',  
    'fastparquet>=2023.10',
    'pyahocorasick>=2.0',
    'en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl'
]

[project.optional-dependencies]
dev = [
    'pytest>=7.4',
    'black>=23.0',
    'ruff>=0.1',
    'isort>=5.12',
    'pre-commit>=3.5',
]
notebooks = [
    'jupyter',
    'ipykernel',
    'matplotlib',
    'seaborn'
]

[tool.setuptools.packages.find]
where = ['.']
include = ['src*']

[tool.black]
line-length = 88
skip-string-normalization = true

[tool.isort]
profile = 'black'

[tool.ruff]
line-length = 88